Add files via upload
I do not know why the game does not pass the load screen when it sets "load flags", and has the respective flags. Is there a way to know specifically why the charge error jumps? It's that it pulls me off the charging screen without telling me why.

If somebody know the error and can solve (=